Split manager module into parts
