+++ crates/rt/src/manager/tests.rs
@@ -50,18 +50,15 @@ fn instantiating_workflow() {
         .unwrap();
 
     let state = manager.state.lock().unwrap();
-    let persisted = &state.committed.workflows[&handle.id()];
+    let persisted = &state.workflows[&handle.id()];
     assert_eq!(persisted.definition_id, "test:latest");
     let orders_id = handle.ids().channel_ids.inbound["orders"];
     assert_eq!(
-        state.committed.channels[&orders_id].receiver_workflow_id,
+        state.channels[&orders_id].receiver_workflow_id,
         Some(handle.id())
     );
     let traces_id = handle.ids().channel_ids.outbound["traces"];
-    assert_eq!(
-        state.committed.channels[&traces_id].receiver_workflow_id,
-        None
-    );
+    assert_eq!(state.channels[&traces_id].receiver_workflow_id, None);
     drop(state); // in order for `test_initializing_workflow()` not to dead-lock
 
     test_initializing_workflow(&manager, handle.ids());
@@ -75,7 +72,7 @@ fn test_initializing_workflow(manager: &WorkflowManager, handle: &WorkflowAndCha
 
     let traces_id = handle.channel_ids.outbound["traces"];
     let state = manager.state.lock().unwrap();
-    let traces = &state.committed.channels[&traces_id].messages;
+    let traces = &state.channels[&traces_id].messages;
     assert_eq!(traces.len(), 1);
     assert_eq!(traces[0].as_ref(), b"trace #1");
 }
@@ -119,7 +116,7 @@ fn sending_message_to_workflow() {
 
     {
         let state = manager.state.lock().unwrap();
-        let orders = &state.committed.channels[&orders_id].messages;
+        let orders = &state.channels[&orders_id].messages;
         assert_eq!(orders.len(), 1);
         assert_eq!(orders[0].as_ref(), b"order #1");
     }
@@ -139,7 +136,7 @@ fn sending_message_to_workflow() {
     );
 
     let state = manager.state.lock().unwrap();
-    let orders = &state.committed.channels[&orders_id].messages;
+    let orders = &state.channels[&orders_id].messages;
     assert!(orders.is_empty());
 }
